+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Batch `Set()`, `Delete()`, and `Close()` may now error (@erikgrinaker)
 
+- `Iterator.Close()` may now error (@erikgrinaker)
+
 - The `SetDeleter` interface has been removed (@erikgrinaker)
 
 ## 0.5.1